rename login and registration files
